feat(users): 🎸 added logic to approve or reject and applicant
Admins can now approve or reject an applicant. I also added the
ion-skeleton to show activity on the page when the server is fetching
for the applicant. I am stil going to add tooltips to the buttons to
show the user what the purpose of the static buttons are.

BREAKING CHANGE: 🧨 none

✅ Closes: https://favr.atlassian.net/browse/GIGWERK-132